ICA fails if n_components=None
If I understand the documentation correctly, you should be able to do ICA with both max_pca_components and n_components set to None.

Currently, I get an error when n_components is None - this is because the loop responsible for selecting components and logging them checks if n_components is a float, and ```else``` (assuming it to be an int I guess) it tries to log it as a string, which will fail if n_components is None. Also, the logger currently tries to log pca.components_, which can be an array. I think it should log len(pca.components_).

If this is not the intended behavior, this PR fixes it. Including these changes, I can run ICA with n_components=None.